+++ src/client/components/ProjectSidebar.tsx
@@ -121,6 +121,9 @@ const style: ThemeUIStyleObject = {
     borderRadius: "100%",
     mr: 2
   },
+  deviationIcon: {
+    ml: "15px"
+  },
   unassignedColor: {
     border: "1px solid",
     borderColor: "gray.3"
@@ -147,6 +150,7 @@ const ProjectSidebar = ({
   lockedDistricts,
   hoveredDistrictId,
   saving,
+  avgPopulation,
   isReadOnly
 }: {
   readonly project?: IProject;
@@ -159,6 +163,7 @@ const ProjectSidebar = ({
   readonly lockedDistricts: LockedDistricts;
   readonly hoveredDistrictId: number | null;
   readonly saving: SavingState;
+  readonly avgPopulation?: number;
   readonly isReadOnly: boolean;
 } & LoadingProps) => {
   return (
@@ -292,7 +297,8 @@ const SidebarRow = memo(
     districtId,
     isDistrictLocked,
     isDistrictHovered,
-    isReadOnly
+    isReadOnly,
+    popDeviationThreshold
   }: {
     readonly district: DistrictGeoJSON;
     readonly selected: boolean;
@@ -304,6 +310,7 @@ const SidebarRow = memo(
     readonly isDistrictLocked?: boolean;
     readonly isDistrictHovered: boolean;
     readonly isReadOnly: boolean;
+    readonly popDeviationThreshold: number;
   }) => {
     const selectedDifference = selectedPopulationDifference || 0;
     const showPopulationChange = selectedDifference !== 0;
@@ -319,6 +326,7 @@ const SidebarRow = memo(
     const deviationDisplay = `${intermediateDeviation > 0 ? "+" : ""}${Math.round(
       intermediateDeviation
     ).toLocaleString()}`;
+
     const compactnessDisplay =
       districtId === 0 ? (
         <span sx={style.blankValue}>{BLANK_VALUE}</span>
@@ -390,7 +398,17 @@ const SidebarRow = memo(
           {populationDisplay}
         </Styled.td>
         <Styled.td sx={{ ...style.td, ...style.number, ...{ color: textColor } }}>
-          {deviationDisplay}
+          <span>{deviationDisplay}</span>
+          <span sx={style.deviationIcon}>
+            {(districtId === 0 && Math.abs(intermediateDeviation) === 0) ||
+            (districtId !== 0 && Math.abs(intermediateDeviation) <= popDeviationThreshold) ? (
+              <Icon name="circle-check-solid" color="#388a64" />
+            ) : intermediateDeviation < 0 ? (
+              <Icon name="arrow-circle-down-solid" color="gray.4" />
+            ) : (
+              <Icon name="arrow-circle-up-solid" color="#000000" />
+            )}
+          </span>
         </Styled.td>
         <Styled.td sx={style.td}>
           <Tooltip
@@ -588,6 +606,7 @@ const SidebarRows = ({
             isDistrictHovered={districtId === hoveredDistrictId}
             districtId={districtId}
             isReadOnly={isReadOnly}
+            popDeviationThreshold={averagePopulation * (project.populationDeviation / 100)}
             votingIds={votingIds}
           />
         );