+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add configurable population deviation [#762](https://github.com/PublicMapping/districtbuilder/pull/762)
+- Display target population symbols [#720](https://github.com/PublicMapping/districtbuilder/issues/720)
 
 
 ### Changed